+++ src/Chaca.ts
@@ -1,4 +1,4 @@
-import { KeyField, KeyFieldProps } from "./core/fields/core/key/key-field";
+﻿import { KeyField, KeyFieldProps } from "./core/fields/core/key/key-field";
 import { ChacaUtils } from "./core/utils";
 import { SchemaInput } from "./core/schema/interfaces/schema";
 import { DatasetSchema } from "./core/dataset-resolver/interfaces/dataset-schema";
@@ -128,7 +128,7 @@ export class Chaca {
    * @param data Data you want to export
    * @param config.filename file name
    * @param config.location location of the file
-   * @param config.format file extension (`'java'` | `'csv'` | `'typescript'` | `'json'` | `'javascript'` | `'yaml'` | `'postgresql'` | `'python'`)
+   * @param config.format file extension (`'java'` | `'csv'` | `'typescript'` | `'json'` | `'javascript'` | `'yaml'` | `'postgresql'` | `'sqlite'` | `'python'`)
    *
    * @example
    * const data = [
@@ -209,7 +209,7 @@ export class Chaca {
    *
    * @param data Data to transform
    * @param props.filename name for the file
-   * @param props.format file extension (`'java'` | `'csv'` | `'typescript'` | `'json'` | `'javascript'` | `'yaml'` | `'postgresql'` | `'python'`)
+   * @param props.format file extension (`'java'` | `'csv'` | `'typescript'` | `'json'` | `'javascript'` | `'yaml'` | `'postgresql'` | `'sqlite'` | `'python'`)
    */
   transform(data: any, props: DumpConfig): DumpFile[] {
     const filter = new GeneratorFilter(this.utils);