+++ src/bin/chaca.ts
@@ -8,6 +8,7 @@ import { js } from "./js";
 import { java } from "./java";
 import { ts } from "./ts";
 import { postgresql } from "./postgresql";
+import { sqlite } from "./sqlite";
 import { python } from "./python";
 
 yargs.command(json);
@@ -17,6 +18,7 @@ yargs.command(js);
 yargs.command(java);
 yargs.command(ts);
 yargs.command(postgresql);
+yargs.command(sqlite);
 yargs.command(python);
 
 yargs.parse();